refactor(ResourceResolver): resolver 객체를 두어 전체 처리 위임 및 폴더 자동 index 검색 삭제
